requirements.txt
flask
python-dotenv
flask-sqlalchemy
psycopg2-binary
flask_cors
cloudinary